requirements.txt
openpyxl-3.1.2
openai-0.28.0
streamlit-1.26.0
nltk-3.8.1
xlsxwriter-3.1.6